# back/docs/swagger.yaml
definitions:
...
  sse.SSESlotUpdateMessage:
    properties:
      endsAt:
        type: string
      id:
        type: string
      isValidated:
        type: boolean
      startsAt:
        type: string
    type: object
info:
  contact:
    email: contact@zide.fr
  description: SlotFinder API Doc
  title: SlotFinder API
  version: 1.0b
paths:
...
  /v1/events/{eventId}/sse:
    get:
      description: Establishes a Server-Sent Events connection to receive real-time
        updates for a specific event
      parameters:
      - description: Event ID
        in: path
        name: eventId
        required: true
        type: string
      responses:
        "200":
          description: Stream of slot update arrays
          schema:
            items:
              $ref: '#/definitions/sse.SSESlotUpdateMessage'
            type: array
        "400":
          description: Invalid event ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Connect to SSE for event updates
      tags:
      - SSE
securityDefinitions:
  AccessTokenCookie:
    description: Enter your bearer token in the format **Bearer &lt;token&gt;**
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"